SAL-30, SAL-31 The changes in r31029 & r31038 were causing the RefApp and Atlassian Gadgets to fail at startup once we updated to the new SAL 2.0 snapshot.  The problem is that sal-core is not actually deployed as an OSGi bundle in the Reference App.
I tried adding com.atlassian.sal.core.util as a private package, which results in the Maven bundle plugin trying to copy that package into the bundle.  Unfortunately, this wanted to drag in a bunch of Apache HttpClient stuff along with it, due to dependencies in URIUtil.  To get unblocked, I just changed RefImplUserManager so that it no longer depends on Assert.  There's probably a better long-term way to do this, maybe by creating a sal-refimpl-core-plugin that packages sal-core as an OSGi bundle with the proper exports.


git-svn-id: https://studio.atlassian.com/svn/SAL/branches/2.0-stable@31095 2c54a935-e501-0410-bc05-97a93f6bca70